Extract relational cancellation and parameters
