+++ crates/truthdb-core/src/engine/relational/prelude.rs
@@ -1,9 +1,9 @@
 #![allow(unused_imports)]
 
 pub(super) use super::{
-    aggregate, api::*, batch::*, collation, constraints::*, context::*, ddl::*, describe::*,
-    dispatch::*, dml::*, hash, helpers::*, lock_analysis::*, plan, procedural::*, query::*,
-    sys_views::*, transaction::*, triggers::*, value,
+    aggregate, api::*, batch::*, cancel::*, collation, constraints::*, context::*, ddl::*,
+    describe::*, dispatch::*, dml::*, hash, helpers::*, lock_analysis::*, parameters::*, plan,
+    procedural::*, query::*, sys_views::*, transaction::*, triggers::*, value,
 };
 
 pub(super) use truthdb_sql::ast::{
